refactor: split ProjectSetupPanel into smaller components
- Created shared types in types.ts
- Split PathInput into its own component
- Created GitUrlInput component
- Created ActionButtons component
- Extracted logic into useProjectSetup hook
- Updated main ProjectSetupPanel to use new components
